services/platform-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.10.3
pydantic-settings==2.6.1
PyJWT==2.10.0
requests==2.32.3
python-multipart==0.0.19


